Add :program option to render/pod2man methods
Currently, render/pod2man use the basename of the $*PROGRAM object
as the name for the generated manual page. This may not be desired
behavior in some situations, like when the executable file does not
share the same name as the file containing the pod. This commit adds
the ':program' parameter to the render/pod2man, allowing the user to
override the manpage program name.