Determine manpage name from pod contents
This commit adds functionality to Pod::To::Man that allows it
to automatically determine the name of the generated manpage from the
contents of its pod. It does this by scanning the pod for
'=head1 NAME' or '=NAME' blocks and extracting the program name from
them. If it cannot determine the program name from the pod, it will
default to the value of $*PROGRAM.basename.